requirements.txt
discord.py[voice]==2.2.3
requests==2.29.0
gitpython==3.1.31
beautifulsoup4==4.12.2
lxml==4.9.2
